Reorganise classes into separate files
