+++ lib/we_transfer_client/future_transfer.rb
@@ -1,0 +1,9 @@
+class FutureTransfer < Ks.strict(:name, :description, :items)
+  def to_create_transfer_params
+    {
+      name: name,
+      description: description,
+      items: items.map(&:to_item_request_params),
+    }
+  end
+end
