+++ lib/we_transfer_client/remote_item.rb
@@ -1,0 +1,2 @@
+class RemoteItem < Ks.strict(:id, :local_identifier, :content_identifier, :name, :size, :mime_type)
+end
