+++ lib/we_transfer_client/remote_transfer.rb
@@ -1,0 +1,2 @@
+class RemoteTransfer < Ks.strict(:id, :version_identifier, :state, :shortened_url, :name, :description, :size, :items)
+end
