+++ spec/integration_spec.rb
@@ -33,7 +33,7 @@ describe WeTransferClient do
       expect(add_result).to eq(true)
     end
 
-    expect(transfer).to be_kind_of(WeTransferClient::RemoteTransfer)
+    expect(transfer).to be_kind_of(RemoteTransfer)
     expect(transfer.id).to be_kind_of(String)
 
     # expect(transfer.version_identifier).to be_kind_of(String)
@@ -44,7 +44,7 @@ describe WeTransferClient do
     expect(transfer.items.length).to eq(3)
 
     item = transfer.items.first
-    expect(item).to be_kind_of(WeTransferClient::RemoteItem)
+    expect(item).to be_kind_of(RemoteItem)
 
     expect(transfer.shortened_url).to be_kind_of(String)
     response = Faraday.get(transfer.shortened_url)
